update ft_utils & makefile

+++ software/ft_utils.c
@@ -19,7 +19,7 @@ FT_STATUS FTUTIL_List()
 		printf("FTDI library version = 0x%x\n", (unsigned int)libraryVersion);
 	else
 	{
-		printf("ERROR FTUTIL_List - Failed reading FTDI library version : FT_GetLibraryVersion(%d)\n", (int)ftStatus);
+		printf("ERROR FTUTIL_List - Failed reading FTDI library version : FT_GetLibraryVersion(%u)\n", ftStatus);
 		return ftStatus;
 	}
 	for(DWORD i = 0; i<FTUTIL_MAX_DEVICES-1; i++)
@@ -29,7 +29,7 @@ FT_STATUS FTUTIL_List()
 		printf("FTDI devices count = %d\n", FTUTIL_List_DevicesCount);
 	else
 	{
-		printf("ERROR FTUTIL_List - Failed to get serial numbers FT_ListDevices(%d)\n", (int)ftStatus);
+		printf("ERROR FTUTIL_List - Failed to get serial numbers FT_ListDevices(%u)\n", ftStatus);
 		return ftStatus;
 	}
 	for(DWORD i = 0; i<FTUTIL_MAX_DEVICES-1; i++)
@@ -42,15 +42,15 @@ FT_STATUS FTUTIL_List()
 	}
 	else
 	{
-		printf("ERROR FTUTIL_List - Failed to get  descriptions FT_ListDevices(%d)\n", (int)ftStatus);
+		printf("ERROR FTUTIL_List - Failed to get  descriptions FT_ListDevices(%u)\n", ftStatus);
 		return ftStatus;
 	}
 	return FT_OK;
 	/*ftStatus = FT_CreateDeviceInfoList(&iNumDevs);
-	if(ftStatus != FT_OK) { printf("Error: FT_CreateDeviceInfoList(%d)\n", (int)ftStatus); return 1; }
+	if(ftStatus != FT_OK) { printf("Error: FT_CreateDeviceInfoList(%u)\n", ftStatus); return 1; }
 	else printf("FT_CreateDeviceInfoList iNumDevs=%d\n", iNumDevs);*/
 	/*ftStatus = FT_GetDeviceInfoList(&DevInfoList, &iNumDevs);
-	if (ftStatus != FT_OK) { printf("Error: FT_GetDeviceInfoList(%d)\n", (int)ftStatus); return 1; }
+	if (ftStatus != FT_OK) { printf("Error: FT_GetDeviceInfoList(%u)\n", ftStatus); return 1; }
 	else {
 		printf("FT_GetDeviceInfoList iNumDevs=%d\n", iNumDevs);
 		for(i = 0; i < iNumDevs; i++) {
@@ -65,7 +65,7 @@ FT_STATUS FTUTIL_List()
 		}
 	}*/
 	/*ftStatus = FT_ListDevices(&iNumDevs, NULL, FT_LIST_NUMBER_ONLY);
-	if(ftStatus != FT_OK) { printf("Error: FT_ListDevices(%d)\n", (int)ftStatus); return 1; }
+	if(ftStatus != FT_OK) { printf("Error: FT_ListDevices(%u)\n", ftStatus); return 1; }
 	else printf("FT_ListDevices iNumDevs=%d\n", iNumDevs);*/
 }
 
@@ -76,22 +76,22 @@ FT_STATUS FTUTIL_Open(int deviceListIndex, ULONG baudRate, UCHAR wordLength, UCH
 		printf("FT_Open(#%d) handle=0x%p\n", deviceListIndex, pHandle);
 	else
 	{
-		printf("ERROR FTUTIL_Open - Failed to open device #%d : FT_Open(%d)\n", deviceListIndex, ftStatus);
+		printf("ERROR FTUTIL_Open - Failed to open device #%d : FT_Open(%u)\n", deviceListIndex, ftStatus);
 		return ftStatus;
 	}
 	//FT_ResetDevice(ftHandle); TRY
 	if ( (ftStatus = FT_SetBaudRate(*pHandle, baudRate)) == FT_OK)
-		printf("FT_SetBaudRate(%d)\n", baudRate);
+		printf("FT_SetBaudRate(%u)\n", baudRate);
 	else
 	{
-		printf("ERROR FTUTIL_Open - Failed to set baud rate : FT_SetBaudRate(%d)\n", (int)ftStatus);
+		printf("ERROR FTUTIL_Open - Failed to set baud rate : FT_SetBaudRate(%u)\n", ftStatus);
 		return ftStatus;
 	}
 	if ( (ftStatus = FT_SetDataCharacteristics(*pHandle, wordLength, stopBits, parity)) == FT_OK)
 		printf("FT_SetDataCharacteristics(BITS:%d STOP_BITS:%d PARITY:%d)\n", wordLength, stopBits, parity);
 	else
 	{
-		printf("ERROR FTUTIL_Open - Failed to configure : FT_SetDataCharacteristics(%d)\n", (int)ftStatus);
+		printf("ERROR FTUTIL_Open - Failed to configure : FT_SetDataCharacteristics(%u)\n", ftStatus);
 		return ftStatus;
 	}
 	//FT_SetUSBParameters
@@ -100,7 +100,7 @@ FT_STATUS FTUTIL_Open(int deviceListIndex, ULONG baudRate, UCHAR wordLength, UCH
 		printf("FT_GetLatencyTimer = %d\n", Latency);
 	else
 	{
-		printf("ERROR FTUTIL_Open - Failed FT_GetLatencyTimer(%d)\n", (int)ftStatus);
+		printf("ERROR FTUTIL_Open - Failed FT_GetLatencyTimer(%u)\n", ftStatus);
 		return ftStatus;
 	}
 	return FT_OK;
@@ -111,9 +111,9 @@ FT_STATUS FTUTIL_WriteData(FT_HANDLE ftHandle, LPVOID transmitData, DWORD bytesT
 	DWORD dwBytesWritten = 0;
 	FT_STATUS ftStatus = FT_Write(ftHandle, transmitData, bytesToWrite, &dwBytesWritten);
 	if (ftStatus != FT_OK)
-		printf("ERROR FTUTIL_WriteData - FT_Write(%d)\n", (int)ftStatus);
+		printf("ERROR FTUTIL_WriteData - FT_Write(%u)\n", ftStatus);
 	if (dwBytesWritten != bytesToWrite) {
-		printf("ERROR FTUTIL_WriteData - FT_Write only wrote %d (of %d) bytes\n", (int)dwBytesWritten,  bytesToWrite);
+		printf("ERROR FTUTIL_WriteData - FT_Write only wrote %u (of %u) bytes\n", dwBytesWritten,  bytesToWrite);
 		ftStatus = FT_OTHER_ERROR;
 	}
 	//usleep(1000);
@@ -125,7 +125,7 @@ FT_STATUS FTUTIL_WaitData(FT_HANDLE ftHandle, DWORD bytesWait, int exactlyBytes,
 	int counter = 0;
 	*bytesAvailable = 0;
 	if (verbose)
-		printf("Waiting data (%d)... \n", bytesWait);
+		printf("Waiting data (%u)... \n", bytesWait);
 	if ( bytesWait==0 )
 		return FT_OK;
 	DWORD bytes = 0;
@@ -134,13 +134,13 @@ FT_STATUS FTUTIL_WaitData(FT_HANDLE ftHandle, DWORD bytesWait, int exactlyBytes,
 		//DWORD temp = 0;
 		FT_STATUS ftStatus = FT_GetQueueStatus(ftHandle, &bytes);
 		if ( ftStatus!=FT_OK ) {
-			printf("ERROR FTUTIL_WaitData - FT_GetQueueStatus - Failed to get number of bytes available to read (error %d)\n", ftStatus);
+			printf("ERROR FTUTIL_WaitData - FT_GetQueueStatus - Failed to get number of bytes available to read (error %u)\n", ftStatus);
 			*bytesAvailable = bytes;
 			return ftStatus;
 		}
 		counter++;
 		if ( (timeout>=0) && (counter*10>timeout) ) {
-			printf("ERROR FTUTIL_WaitData - Waiting timeout expired %d ms. Available bytes = %d\n", timeout, bytes);
+			printf("ERROR FTUTIL_WaitData - Waiting timeout expired %d ms. Available bytes = %u\n", timeout, bytes);
 			*bytesAvailable = bytes;
 			return FT_OTHER_ERROR;
 		}
@@ -156,12 +156,12 @@ FT_STATUS FTUTIL_WaitData(FT_HANDLE ftHandle, DWORD bytesWait, int exactlyBytes,
 	//	printf("\n");
 	if (exactlyBytes) {
 		if (bytes!=bytesWait) {
-			printf("ERROR FTUTIL_WaitData - Available invalid number of bytes %d (expected %d)\n", bytes, bytesWait);
+			printf("ERROR FTUTIL_WaitData - Available invalid number of bytes %u (expected %u)\n", bytes, bytesWait);
 			return FT_OTHER_ERROR;
 		}
 	}
 	else if (bytes<bytesWait) {
-		printf("ERROR FTUTIL_WaitData - Available invalid number of bytes %d (expected %d)\n", bytes, bytesWait);
+		printf("ERROR FTUTIL_WaitData - Available invalid number of bytes %u (expected %u)\n", bytes, bytesWait);
 		return FT_OTHER_ERROR;
 	}
 	return FT_OK;
@@ -172,11 +172,11 @@ FT_STATUS FTUTIL_ReadData(FT_HANDLE ftHandle, LPVOID readData, DWORD bytesToRead
 	*bytesReaded = 0;
 	FT_STATUS ftStatus = FT_Read(ftHandle, readData, bytesToRead, bytesReaded);
 	if (ftStatus!=FT_OK) {
-		printf("ERROR FTUTIL_ReadData - FT_Read(%d)\n", ftStatus);
+		printf("ERROR FTUTIL_ReadData - FT_Read(%u)\n", ftStatus);
 		return ftStatus;
 	}
 	if ( *bytesReaded != bytesToRead ){
-		printf("ERROR FTUTIL_ReadData - Readed invalid number of bytes %d (expected %d)\n", *bytesReaded, bytesToRead);
+		printf("ERROR FTUTIL_ReadData - Readed invalid number of bytes %u (expected %u)\n", *bytesReaded, bytesToRead);
 		return FT_OTHER_ERROR;
 	}
 	return FT_OK;
